requirements.txt
pynput
python-socketio
python-engineio
